+++ epi/src/lib.rs
@@ -233,6 +233,12 @@ pub struct NativeOptions {
     /// The initial size of the native window in points (logical pixels).
     pub initial_window_size: Option<egui::Vec2>,
 
+    /// The minimum window size
+    pub min_window_size: Option<egui::Vec2>,
+
+    /// The maximum window size
+    pub max_window_size: Option<egui::Vec2>,
+
     /// Should the app window be resizable?
     pub resizable: bool,
 
@@ -251,6 +257,8 @@ impl Default for NativeOptions {
             drag_and_drop_support: false,
             icon_data: None,
             initial_window_size: None,
+            min_window_size: None,
+            max_window_size: None,
             resizable: true,
             transparent: false,
         }